requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
sentence-transformers
openpyxl
boto3